test(e2e): split node-served smoke spec from offline liveness tier
The previous commit rewrote production-liveness.spec.ts for the
node-served webapp (iframe loader, title "Freenet", live onboarding). But
the default playwright config (testDir ".", ignores only node-e2e/) runs
that spec in CI's ui-playwright-tests against the OFFLINE `vite preview`
build (title "Raven", direct mount, mock posts) — so the rewrite broke CI.

The two harnesses need opposite assertions:
- offline tier (default config, vite preview):  "Raven", direct DOM, mocks
- node-served smoke (scripts/smoke-test-production.sh): "Freenet", iframe#app,
  live-network onboarding

Fix:
- restore web/tests/production-liveness.spec.ts to the original offline
  assertions (what CI ui-playwright-tests runs across chromium/firefox/webkit).
- move the node-served assertions to web/tests/node-e2e/production-liveness.spec.ts
  — the default config ignores node-e2e/, so CI's offline job never runs it.
- point scripts/smoke-test-production.sh at the node spec via
  playwright.node.config.ts (chromium-only, no self-booted webServer; the
  smoke runner supplies the already-live BASE_URL).

Verified: offline spec lists under default config (3 browsers); node spec
lists under node config (chromium); smoke runner 2/2 green against the live
v0.1.0 publish.
